+++ example/input/InputSamples.dart
@@ -18,7 +18,7 @@ void main() {
     label.profile.width = "70";
     view.addChild(label);
 
-    var input = type == "multiline" ? new MultilineBox(): new TextBox(null, type);
+    var input = type == "multiline" ? new TextArea(): new TextBox(null, type);
     input.on.change.add((event) {
       TextView inf = input.nextSibling;
       inf.text = input.value;